+++ src/main/java/pokeraidbot/commands/WhatsNewCommand.java
@@ -22,6 +22,8 @@ public class WhatsNewCommand extends ConfigAwareCommand {
         final String message;
         if (config.getLocale().equals(LocaleService.SWEDISH)) {
             message = "**Nytt i " + BotServerMain.version + ":**\n\n" +
+                    "- !raid group titel har nu raidstart och -slut\n" +
+                    "- Fix: Eventlyssnare för grupper rensas undan bättre\n" +
                     "- !raid overview och !raid list är omgjorda, listar nu bland annat grupptider i stället för ETA\n" +
                     "- Gen3 pokemons finns nu inlagda (rapportera gärna om något är fel!)\n" +
                     "- Alias för vissa kommandon, så det blir mindre att skriva - t.ex. " +
@@ -34,6 +36,8 @@ public class WhatsNewCommand extends ConfigAwareCommand {
                     "- Fix: EX raids kommer inte längre skapas via botintegration";
         } else {
             message = "**New in " + BotServerMain.version + ":**\n\n" +
+                    "- Add raid duration to group message header\n" +
+                    "- Fix: Better cleanup of event listeners\n" +
                     "- !raid overview and !raid list remade, for example they now list times for groups instead of ETA\n" +
                     "- Gen3 pokemons added to repository\n" +
                     "- Aliases for commands, to reduce typing during winter - for example " +